+++ packages/adapter-node/index.js
@@ -12,7 +12,7 @@ export default function ({ out = 'build' } = {}) {
 	const adapter = {
 		name: '@sveltejs/adapter-node',
 
-		async adapt(utils) {
+		async adapt({ utils }) {
 			utils.log.minor('Copying assets');
 			const static_directory = join(out, 'assets');
 			utils.copy_client_files(static_directory);