+++ packages/kit/src/runtime/server/parse_body/index.js
@@ -1,11 +1,13 @@
 import { read_only_form_data } from './read_only_form_data.js';
 
-/** @param {import('types/hooks').Incoming} req */
-export function parse_body(req) {
-	const raw = req.rawBody;
+/**
+ * @param {import('types/hooks').StrictBody} raw
+ * @param {import('types/helper').Headers} headers
+ */
+export function parse_body(raw, headers) {
 	if (!raw) return raw;
 
-	const [type, ...directives] = req.headers['content-type'].split(/;\s*/);
+	const [type, ...directives] = headers['content-type'].split(/;\s*/);
 
 	if (typeof raw === 'string') {
 		switch (type) {