+++ packages/kit/types/index.d.ts
@@ -3,7 +3,14 @@
 
 import './ambient-modules';
 
-export { Adapter, AdapterUtils, Config } from './config';
+export { Adapter, AdapterUtils, Config, ValidatedConfig } from './config';
+export { EndpointOutput, RequestHandler } from './endpoint';
 export { ErrorLoad, Load, Page } from './page';
-export { Incoming, GetSession, Handle, ServerResponse as Response, ServerFetch } from './hooks';
-export { ServerRequest as Request, EndpointOutput, RequestHandler } from './endpoint';
+export {
+	Incoming,
+	GetSession,
+	Handle,
+	ServerRequest as Request,
+	ServerResponse as Response,
+	ServerFetch
+} from './hooks';