+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sveltejs/adapter-vercel
 
+## 1.0.0-next.20
+
+### Patch Changes
+
+- c3d36a3: ensure `content-length` limit respected; handle `getRawBody` error(s)
+- 028abd9: Pass validated svelte config to adapter adapt function
+- Updated dependencies [6372690]
+- Updated dependencies [c3d36a3]
+- Updated dependencies [bf77940]
+- Updated dependencies [2172469]
+- Updated dependencies [028abd9]
+  - @sveltejs/kit@1.0.0-next.110
+
+## 1.0.0-next.19
+
+### Patch Changes
+
+- dca4946: Make kit a peerDependency of the adapters
+- Updated dependencies [261ee1c]
+- Updated dependencies [ec156c6]
+- Updated dependencies [586785d]
+  - @sveltejs/kit@1.0.0-next.109
+
+## 1.0.0-next.18
+
+### Patch Changes
+
+- dad93fc: Fix workspace dependencies
+- Updated dependencies [dad93fc]
+- Updated dependencies [37fc04f]
+  - @sveltejs/kit@1.0.0-next.108
+
 ## 1.0.0-next.17
 
 ### Patch Changes